+++ src/main/kotlin/com/bandage/v1/domain/member/model/Member.kt
@@ -1,7 +1,6 @@
 package com.bandage.v1.domain.member.model
 
 import com.bandage.v1.global.common.domain.BaseTimeEntity
-import com.bandage.v1.global.common.domain.enums.MemberRole
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
@@ -15,22 +14,16 @@ import org.hibernate.annotations.SQLRestriction
 @SQLRestriction("deleted_at IS NULL")
 open class Member(
     email: String,
-    password: String,
     name: String,
     contact: String,
-    role: MemberRole = MemberRole.MEMBER,
 ) : BaseTimeEntity() {
     @Id
     @Column(name = "member_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long? = null
+    val id: Long = 0L
 
     @Column(name = "email", unique = true, nullable = false)
     var email: String = email
-        protected set
-
-    @Column(name = "password", nullable = false)
-    var password: String = password
         protected set
 
     @Column(name = "name", nullable = false)
@@ -41,20 +34,14 @@ open class Member(
     var contact: String = contact
         protected set
 
-    @Column(name = "role", nullable = false)
-    var role: MemberRole = role
-        protected set
-
     companion object {
         fun create(
             email: String,
-            password: String,
             name: String,
             contact: String,
         ): Member =
             Member(
                 email = email,
-                password = password,
                 name = name,
                 contact = contact,
             )
