+++ src/main/kotlin/com/bandage/v1/facade/MemberWithdrawFacade.kt
@@ -3,7 +3,6 @@ package com.bandage.v1.facade
 import com.bandage.v1.domain.auth.service.MemberAuthService
 import com.bandage.v1.domain.member.service.MemberService
 import com.bandage.v1.global.security.jwt.RefreshTokenRepository
-import com.bandage.v1.global.util.SecurityUtil
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
 
@@ -14,8 +13,7 @@ class MemberWithdrawFacade(
     private val refreshTokenRepository: RefreshTokenRepository,
 ) {
     @Transactional
-    fun withdrawMember() {
-        val memberId = SecurityUtil.getCurrentMemberId()
+    fun withdrawMember(memberId: Long) {
         memberService.deleteMember(memberId)
         memberAuthService.deleteMemberAuth(memberId)
         refreshTokenRepository.delete(memberId)
