+++ src/main/kotlin/com/bandage/v1/global/async/event/MemberLoginEvent.kt
@@ -1,5 +1,6 @@
 package com.bandage.v1.global.async.event
 
+@Deprecated(message = "회원 인증 로직: 이벤트 기반 아키텍처 적용 X")
 data class MemberLoginEvent(
     val memberId: Long,
     val refreshToken: String,
@@ -7,14 +8,14 @@ data class MemberLoginEvent(
         eventType = EventType.MEMBER_LOGIN,
         aggregateId = memberId.toString(),
     ) {
-    companion object {
-        fun of(
-            memberId: Long,
-            refreshToken: String,
-        ): MemberLoginEvent =
-            MemberLoginEvent(
-                memberId = memberId,
-                refreshToken = refreshToken,
-            )
-    }
+//    companion object {
+//        fun of(
+//            memberId: Long,
+//            refreshToken: String,
+//        ): MemberLoginEvent =
+//            MemberLoginEvent(
+//                memberId = memberId,
+//                refreshToken = refreshToken,
+//            )
+//    }
 }
